backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiofiles==23.2.1
slowapi==0.1.9
email-validator==2.1.0
pandas==2.2.0
openpyxl==3.1.2
